requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
requests==2.32.3
groq==0.9.0
openai==1.35.14
gunicorn==22.0.0
beautifulsoup4==4.12.3
lxml==5.2.2

